CSS, HTML, IMAGES, FONTS, HEAD & DELETED FILES
I changed some CSS so that the frontpage looks better with the human and t-rex. Also i added a font that is will be used to the tag-line.

Changed so background picture so it looks more smooth.

Added the font link in the head include.

Moved the "Popup from search" down to the other searhbar for better structure and overview.

Deleted the index1.html, index_broker.html and example.html.  Also deleted the old background picture from the folder.